Use ps instead of jps to get Buckd PIDs
Summary: Current implementation of SOS report obtains the PIDs of Buckd by grepping and parsing the output of jps. However, when the length of java args is too long, jps cuts off the last part of the args, causing to grep it wrongly. This Diff makes use of ps instead.

Reviewed By: swgillespie

fbshipit-source-id: 58002784f4ef1f8281a6e547179eed2d75e248ce