apple | minor refactoring of AppleBundlePart subclasses
Summary: Removing `getContentHashSourcePath` from the superclass as semantics of the field is different — for `DirectoryAppleBundlePart` and `FileAppleBundlePart` it points to the plain text file with single hash value, for `DirectoryAppleBundlePart` it points to JSON map of relative path of the contained file -> hash of that file. Renaming it in `DirectoryAppleBundlePart` properly.

Reviewed By: milend

fbshipit-source-id: 3ab5eb970d62d67a078607713296be7dbba79dd5