Fix testAppleBinaryWithLinkGroupsWithMixedFilterWithTreeTraversal
Summary:
# Changes

Previously, we were calling `get_value_from_c`, declared in `C/C.h` without actually including the header. Due to lack of declaration, in Xcode 11.x, this was just raining a warning.

In order to access the header, we had to add the dep `B -> C`.

# Breakage Reason

The upgrade to Xcode 12 turned a warning into an error which caused the breakage.

Reviewed By: blackm00n

fbshipit-source-id: 04b4e3b3ce9f3290a5e10ba37d695d9e80a2be5e