Add configuration to control Watchman sync timeout
Summary:
# Changes

We would like to change the default timeouts for Watchman due to flakeyness in test environments. In addition to the command timeout, we would like to change the [sync timeout](https://facebook.github.io/watchman/docs/cmd/query.html#synchronization-timeout-since-21) which is currently hardcoded.

Here, we make two changes:

- Add `project.watchman_sync_timeout_ms` configuration setting that can be used to configure the value
- Rename `WATCHMAN_CLOCK_SYNC_TIMEOUT` -> `DEFAULT_WATCHMAN_CLOCK_SYNC_TIMEOUT_MS` for improved clarity

NOTE: This diff contains *no* behavioral changes at all. It's serves as preparation to the diff that will change the timeouts.

Reviewed By: blackm00n

fbshipit-source-id: ce9ccad287131990a0abbb374d55db4e6a7b4f57