Only have final the CLI command wait on background tasks
Summary:
When buck is configured with `daemon.flush_events_before_exit=true`
parallel command executions are blocked from returning until all commands
finish running. This is not great if a fast command is run while a slower is
running. Imagine a `buck build` concurrently with a fast `buck query`. The
reason this happens is because the `AsyncBackgroundTaskManager` does not start
running any tasks until it has been notified that all commands have ended.
Meaning that commands appear stuck waiting for futures that are not getting
run.

Change the `TaskManagerCommandScope` to only block if there are no other
commands left running. This ensures the final command awaits its tasks, while
the others can finish early.

fbshipit-source-id: 31334c045f83564e7fc9787c0a447205e09a98fc