Fix comp db platform behavior for Apple platforms
Summary:
# The Bug

The `#compilation-database` flavor does not use the same platform fallback behavior at the Apple layer as its underlying code that derives the platform to compute the flags.

This ends up with the target SDK flags missing if an explicit platform flavor is not provided

# The Fix

We already have a function that computes the correct platform with the appropriate fallback behavior (`AppleDescriptions.getAppleCxxPlatformsFlavorDomain()`), so just switch to using it.

The new behavior is guarded by a temporary config option for safety.

Reviewed By: blackm00n

fbshipit-source-id: 2a074a0043b404475f93931188925b6690ff3bc4