eden: resolve redirection symlink on Windows
Summary:
The buck-out directory is a symlink on Windows, but simply walking its parent
will make the is_eden function to claim that it is in an eden mount. What we
actually want is to test whether the pointed to directory is in an edenfs
mount. For that, we need to resolve the symlink.

Now starts the Windows and Python shenanigans. Ideally, we could simply use
os.path.realpath, but that only resolves symlinks on Python3.8, which we do not
have on laptops or Sandcastle. As an alternative, Path.resolve() could be used,
but that's buggy on Python3.6.2 that we use on Sandcastle for testing buck...
Instead, we have to manually test for the symlink and resolve it by hand. The
eden_helpers can use Path.

Reviewed By: fanzeyi

fbshipit-source-id: 5c71aff0bd5e07343459f6376571f45fe43ddbe0