Make `MergedTargetGraph` more memory efficient
Summary:
The previous implementation used a `MutableDirectedGraph` to
build a `DirectedAcyclicGraph`, which meant we kept effectively
two extra copies of the original target graph in memory when
constructing the latter.  Furthermore, the former use a
`LinkedHashMultimap`s under the covers, which are pretty memory
hungry.

This diff re-implements to avoid creating an additional underlying
copy and just generates the necessary graph for the various
methods on the fly.

Additional context:
https://fb.prod.workplace.com/groups/sandcastleqna/permalink/4520559987992581/

Reviewed By: cjhopman

fbshipit-source-id: 086676318acb64e58ceaf455b454be33d850216d