Move downward protocol establishment outside of while(true) loop
Summary:
Second try at moving downward protocol outside of while loop

Last time I moved the protocol out of the loop, the change was reverted because the code would hang at the `readProtocol` line, then closing the named pipe caused an `IOException` (which was actually a `ClosedChannelException` to be thrown, causing an error level message to be printed:

```
LOG.error(e, "Cannot read from named pipe: %s", namedPipeName);
```

This error was shown to the user and spammed the user's console.

When the protocol was printed in the while loop, closing the named pipe causes an info level message to be printed:

```
LOG.info("Named pipe %s is closed", namedPipeName);
```

This log level is not shown in the console by default.

Furthermore, in the outer `catch`, `PipeNotConnectedException` , which is what Windows throws in place of POSIX' `ClosedChannelException`, prints an info error log instead of error log. This change also makes the POSIX log level for this error consistent with the windows version and moves the protocol reading outside of the while loop.

As a side note, the named pipe is now closed after canceling the handler (instead of before), so there should also be fewer instances of `ClosedChannelException` / `PipeNotConnectedException`

Reviewed By: bobyangyf

fbshipit-source-id: 0b79e15b4ea1c4a8dc752aee78820b79bb36494c