DirectedAcyclicGraph: add and use memory efficient builders
Summary:
Previously, creating a `DirectedAcyclicGraph` required creating a
 `MutableDirectedGraph`, then passing it to the former's constructor
where it'd perform copies of the node index and edge multimaps,
which required 2x the size of the original graph to complete.  In
situations with large graphs, this can cause significant memory use
(e.g. fbcode's TD is getting stuck GC'ing due to this copy:
https://www.internalfb.com/intern/diff/D24353308/?dest_fbid=2906789779542824&transaction_id=377674397010078).

This diff adds two builder methods which avoid copying the sets/maps
when constructing the graph and so can be used in more memory
constrained environment.

Reviewed By: bobyangyf

fbshipit-source-id: af164199bdf83d8659cac2c4f05eaf5b2a1e7e80