Re-purpose `cxx.use_archives` platform settings
Summary:
The `cxx.use_archives` platform config settings was added
in D24376485 to make all libraries built for a given platform
to avoid using archives.  It turned out, however, that this
all-or-nothing approach likely isn't desirable, as there's linker
performance regressions when switching it on due to the
increase in input files the linker has to process.  As such,
we switched gears in D24427600 to allow opting out of
archives at the library level, when beneficial (e.g. single
object-file libraries).

However, some platforms don't support linking without
archives, so this diff re-purposes the `cxx.use_archives` and
turns it into a config that allows platforms to say whether they
support archive-less linking or not.

Reviewed By: asp2insp

fbshipit-source-id: d29863832cc3d283f0c14ac09162eae04c999fb0