Refactor windows named pipe input stream
Summary:
Currently getting a new `NamedPipeInputStream` instance requires creating a new `WindowsNamedPipeClientReader`.

We should be able to get a `NamedPipeInputStream` without creating a `WindowsNamedPipeClientReader`. After closing the input stream, the reader can be reused if it's going to read from the same named pipe (i.e. we pass the same named pipe handle into the `NamedPipeInputStream`'s constructor). We need a new instance of the input stream to continue to read from the same named pipe, not an entire new instance of the reader itself.

Lifted `NamedPipeInputStream` out of `WindowsNamedPipeClientBase` because it isn't only used by the client. The server can use this (and it does use it in the handler) too

`CreateFile` doesn't have `link` in `WindowsNamedPipeClientBase`'s javadoc because I'm getting an error similar to this one in SC's ant build: https://bugs.eclipse.org/bugs/show_bug.cgi?id=380879
  Tag link: can't find CreateFile
Follow-up diff will refactor windows named pipe output stream in a similar way

Reviewed By: mykola-semko

fbshipit-source-id: 8ad4abc69e67f884073b43b0c86daac10cb2d3d7