Add option to enable SHA1 logging
Summary:
This adds a new option, `build.sha1_logging_threshold`, which can be used to enable logging of SHA1 hashing operation for files. This is an essential tool in analysing build performance and debugging.

If `build.sha1_logging_threshold` is unset, then logging if disabled. Otherwise, it defines the threshold, in microseconds, so that any hash operations that take at least that amount of time will be logged. Setting it to `0` effectively enables logging for all SHA1 hash operations.

Reviewed By: blackm00n

fbshipit-source-id: e70cb30236116a24e7553b16ff94dc2361859ec3