apple | shorten the names of hashing targets
Summary: When building on linux with incremental bundling enabled it's possible to hit the file name length constraint of 255 symbols. Long file name is for the file hashing target output directory. In this diff I'm adding an option to omit the file name of hashed file from such target name.

Reviewed By: milend

fbshipit-source-id: f80d3683f4a96924b8c7eb27c7b360487a520016