Use classpath arg file when running tests with JDK >= 9
Summary:
Looking at D15966552 when this was introduced, we see:
`However, as of yet, Buck doesn't always know when it's launching Java 9+, so we need the old approach to handle when Buck launches a target that targets Java 8-, but java on the path turns out to be 9+`

Now that we specify a `java_for_tests_version`, we can also use that to know when we should be writing the classpath args file. That is what this diff does.

Reviewed By: jselbo

fbshipit-source-id: b097f1802ce0dd555c37e69250fae9d37f449136