Add opt-in to deduplicate omnibus roots
Summary:
With omnibus linking, each top-level binary rule (e.g. Python binaries
and tests) performs its own omnibus link, each consisting of a single,
large merged omnibus library link and several smaller "root" links for
(e.g. for C++ python extensions).  As such, driven by the presence of
common "roots" in the build (i.e. roots which linked into several binaries),
the amount of omnibus root links if approximately
`O(<# potential roots in build> * <# binaries>)`,
which scales really poorly as the number of binaries and tests in a
build increases.

As, in many cases, some of the root nodes in different binaries are
actually identical, this diff adds new, opt-in mode which attempts to
which deduplicate roots links in this case.  For example, if all binaries
in a build depend on the same C++ python extension and build it under
the same conditions (e.g. same flags, deps, etc), we can link it just once
and reuse across all binaries, rather than linking it separately for each
binary.

Reviewed By: asp2insp

fbshipit-source-id: 5751b67c827bb4446779df5c1cc84dde9a507b81