Avoid redundant topo-sorting
Summary:
As part of omnibus linking, we walk the topo-sorted native linkable
graph found from each of the roots independently, which can mean
a lot of re-walking when there are lots of roots with common dep
trees.

This diff performs a single topo-sort over the combined root nodes/deps
to avoid this.

Reviewed By: bobyangyf

fbshipit-source-id: ae401e468a5f63e603b5aa1f7c1dd17a31920873