nailgun: no busy looping in select on Windows
Summary:
On Windows, the PeekNamedPipe will as expected immediately return, but this has
the negative side effect of making the `select` function a busy loop,
effectively using a full core, and thus reducing compilation speed.

As a workaround, let's just sleep a bit in select if nothing can be read from
the named pipe.

Reviewed By: stepancheg

fbshipit-source-id: 0d2267ac5da3294d2022ed84cd8499c2fa5db66b