Avoid buffering large global symbol lists in memory
Summary:
Previously, we'd buffer about three copies of the global symbols
lists in memory -- once as a `LinkedHashSet` to collect all symbols,
once as the formatted symbol file lines, and again as the symbol
file contents (the lines concatenated).

This diff replaces this with a more memory efficient implementation,
which requires only a single copy (as a less memory-hungry `HashSet`).

fbshipit-source-id: a1707f11634e25434cbb1197ef704c0da156a017