Refactor BuckAddDependencyIntention creation
Summary:
In a following diff, we will add a separate intention that handles the case when the dep is a library.

While it is possible to do this without adding new files, it would make `BuckAddDependencyIntention` much harder to read and make changes.

Moved all the creation logic into a factory, and added an abstract base class.

Reviewed By: cute-jumper

fbshipit-source-id: 79a6c7faef2b6b5a41335fdb20164e5ff16b59b9