Androidx target transformer
Summary:
For context, when deduce a Buck target from a class in a library file, we can only get the target that directly contains the jar.

For androidx libraries, we should avoid this and use other targets that were meant for including as a dep.

Also added accessors to the mapping in D24877766.

Reviewed By: cute-jumper

fbshipit-source-id: ca2440f94bff4e39dd96dab666379247233bad06