Handle missing library case for autodeps
Summary:
Before, we could only provide quickfixes for including Buck deps that belonged to modules.

This handles the case where the class belongs to a library, in short we find the library and the Buck target associated to the library, and update the IntelliJ and Buck dependencies respectively.

A current limitation is that we find the library target that directly contains the class, and not a parent target that might be more commonly used. (e.g. `androidx`)

Reviewed By: cute-jumper

fbshipit-source-id: b143dc7d289b1e7e9d19f01118cce05e4669fe45