Fix creation of provisioning profile metadata
Summary:
Provisioning profiles can have the app ID specified using qualified or non-qualified keys: `application-identifier` or `com.apple.application-identifier`.

Most provisioning profiles contain both keys but *some* contain only the fully qualified one. This ended up in a `NullPointerException` as the code always assumed the non-qualified one would be present which is not the case.

Here, we update the code, so that we look for both keys.

Reviewed By: blackm00n

fbshipit-source-id: 306bb9599eaec9d970448522aa2d75c6450efec3