Fix crash when adding embedded bundle given IB is on
Summary:
Files and directories from embedded bundles when incremental bundling is on are not copied directly to "apple bundle destination" directories. E.g. for the following result bundle structure:
```
└── Contents
    ├── Frameworks
    │   └── TestFramework.framework
    │       ├── Resources
    │       │   ├── Info.plist
    │       │   └── PkgInfo
    │       └── TestFramework
    ...
```
not the whole `Contents/Frameworks/TestFramework.framework` directory will be copied when IB is on, but separate `Contents/Frameworks/TestFramework.framework/Resources/Info.plist` will be copied. So in this case we not only need to create "apple bundle destination" which is `Contents/Frameworks`, but `Contents/Frameworks/TestFramework.framework/Resources` as well.

Reviewed By: milend

fbshipit-source-id: 7b7dae8b3cf7d193b3bfbc8dd5478369513a013f