Use a real target for dummy omnibus deduplication
Summary:
D24876462 generated a synthetic target to deduplication dummy
omnibus links across binaries/tests in the build, causing some issue
due to no corresponding target node (and also fixes an issue with
the C++ platform not being used as a flavor).

This diff switches to use a real target provided by the user.

Reviewed By: asp2insp

fbshipit-source-id: 425a9d95e3b1977f2d0c84a7a288355c13e05b80