Allow to set both SourcePath and RelPath params for JavacPluginProperties
Summary:
Previously it was just SourcePath used to pass path to compiled signatures file.
But we'll also need to pass the output directory, so that Nullsafe plugin can write a report to it.

Therefore here we're extracting path params into a separate class and extending it to contain both SourcePaths and RelPaths.

Reviewed By: mityal

fbshipit-source-id: b4825c559acb78cdc661738deffe9aec2cc7b478