Avoid prematurely adding Java library rules to build graph for flavored targets
Summary:
See Nullsafe's javadoc for explanation of integration.

The problem was that `DefaultJavaLibraryRules.buildLibrary()` not just built
the library but as a side effect added this library to the build graph.  So,
returning `Nullsafe` build rule from Java/AndroidLibrary description didn't
have an effect since the slot in the build graph was already occupied.

Graphically:
```
-> enter: JavaLibraryDescription.createBuildRule(target=//:foo#nullsfex)
-| graph: {}
...
    if (Nullsafe.hasSupportedFlavor(flavors)) {
      javaLibrary = defaultJavaLibraryRules.buildLibrary()); // = JavaRule1
      -> enter: buildLibrary()
         ...
      <- exit: buildLibrary()
      -| graph: {//:foo#nullsafex -> JavaRule1}
      return Nullsafe.create(graphBuilder, javaLibrary); // = NullsafeRule1
    }
<- exit: createBuildRule
- try to insert //:foo#nullsafex -> NullsafeRule1 into graph but can't, since the mapping already exists
-| graph: {//:foo#nullsafex -> JavaRule1}

```

We don't want that. We also don't need the extra ABI rules. Therefore we create
and use a new builder method, that doesn't have any effects on the build graph
and build a JavaLibrary without ABIs and whatnot.

Reviewed By: mityal

fbshipit-source-id: c6c13417f1fe2d7cf0bde24c746f033be6c9bcf7