Add SourcePath based params to JavacPluginProperties
Summary:
NullsafeX javac plugin requires configuration params that depend on other build
targets (see next diff for example) or more generaly on some other SourcePaths.

To accomodate for this use-case we can add a params map with SourcePaths, that
are resolved to RelPaths when `ResolvedJavacPluginProperties` are built, and
then resolved against cell root to AbsPaths when ResolvedJavacOptions are
constructed.  This setup is compatible with MBRs and RE.

NOTE: Resolving SourcePaths to RelPaths while constructing
`ResolvedJavacPluginProperties` is a bit premature, but this is a general
problem with java library builds that paths are resolved too eagerly (i.e. at
"construction" rather than "execution" phase).

Reviewed By: mityal

fbshipit-source-id: ea8693f1b29465dc8cefc1a7125fb5d957f26d2e