Implement Nullsafe build rule and #nullsafex JavaLibrary flavor
Summary:
NullsafeX is a javac compiler plugin. For a number of reasons we need to use a flavor to integrate it
with the build and existing infrastructure.

Here we implement a basic Nullsafe build rule that wraps and augments JavaLibrary by enabling and configuring
Nullsafe javac plugin for it.

In the follow-up diffs we're going to extend Nullsafe build rule with extra config options and an
additional mode of operation.

NOTE: NullsafeX and this build rule are in pre-production/experimental stage, and will get changed and improved.

Reviewed By: mityal

fbshipit-source-id: 8bb3e2b2ffec48978b08e8ca7af6088b31092ba5