Ignoring module-info.class
Summary:
Fixing a crash when buck tries to process the module-info.class, which is unnecessary anyway.

This problematic class is added by external JARs that were compiled with Java9. This issue was fixed in the open sourced bug (see https://github.com/facebook/buck/issues/2498), but the fix didn't make its way to our codebase so I'm adding it here.

Reviewed By: lebentle

fbshipit-source-id: 09b92a39530ccd52b99bc3b39364e0c6da930e66