Make resolved Python components lazy
Summary:
Every Python binary/test in a build has to resolve it's transitive
components.  As implemented, this meant copying `O(N * M)`
maps with large paths, which takes up lots of memory.

This diff makes this lazy, so that we only resolve paths when they
are iterated by build steps and avoid building new maps altogether.

Reviewed By: bobyangyf

fbshipit-source-id: c4c65884a6e39ae6dce4a5089c547351141fce58