PexStep: don't buffer component map in memory
Summary:
The current implementation needs to buffer 2x copies of the
transitive component map -- one for the map objects and one
for the serialized JSON string.  In build with large dep trees
and lots of top-level binaries with share many of these deps,
these generate huge amounts of wasted memory (which is
kind of `O(n^2)`).

This diff switches to serializing and streaming the JSON directly
to the PEX process to avoid any in-memory buffering.

Reviewed By: cjhopman

fbshipit-source-id: 215c655157f5ef29a9276ed0537b8eadc34164ad