Do not hold SourcePath data in ResolvedJavacPluginProperties
Summary:
In order to create this class in JavaCD we need to get rid of SourcePath data from it.

`ResolvedJavacOptions.getJavaAnnotationProcessorParams().getPluginProperties()` returns a list of `ResolvedJavacPluginProperties` that should not have any `SourcePath` values.

The only field that holds SourcePath is `inner`, but when I remove it I receive an error such as `javac unable to load annotation processor: <name of the processor>`
In JavaCD process we could set inner to `null` and that should work as this field does not accessed during the execution.

Reviewed By: irenewchen

fbshipit-source-id: 350047f55e947113d1566a18facce216b47f5c3e