Fix multiarch code when linker maps are disabled
Summary:
`MultiarchFile` makes a (wrong) assumption about the internal logic of `CxxLink` and when it produces a linker map. Specifically, if we disable linker maps globally using `cxx.linker_map_enabled=false`, `MultiarchFile` still thinks that a linker map file will be created.

The fix is to just rely on the authoritative state and ask `CxxLink` whether linker maps are enabled or not.

NOTE: `getLinkerMapPath()` has another use which requires the current semantics, so we do not want to change its behaviour.

Reviewed By: blackm00n

fbshipit-source-id: ec2d3c2a12e45122d75aca31c8b0c9ace013275f