Inline most of `CompilerOutputPaths` static methods
Summary:
The main goal is to pass output paths to java compilation instead of creating them in runtime based on a flavored target (ex. with source-abi flavor).

Instead of implicitly creating `CompilerOutputPaths` object and then returning a path value from it, let's inline most of the method to call sites. In the later diffs, we would reuse `CompilerOutputPaths` object created in advance.

Reviewed By: cjhopman

fbshipit-source-id: 9f2e6fc0cd0c1d1daa2f9c2116099b490744409b