Remove BaseBuckPath from JavacExecutionContext
Summary:
With `BaseBuckPaths` removal, the way on how depFilePath is obtained is changed as well. Instead of creating `CompilerOutputPaths` object  `CompilerOutputPaths.getDepFilePath` receives an output jar dir and return a path to a dep file.

As `Invocation` could build classes, source-abi, and source-only-abi artifacts, it must know about 3 possible output path directories. This change is also modifies `Jsr199JavacInvocation` class in a way to receive 3 output pass directories in the constructor.

Reviewed By: irenewchen

fbshipit-source-id: e30df7d43d6e5b508410860aafc4ab8c54b7578a