reduce VFS overlay to 1 per module
Summary:
First draft of an improved VFS overlay approach:
- no VFS overlays for dependent modules
- 1 VFS overlay for mixed modules to mask the public module map with the unextended one that hides the Swift header

Getting the required paths was tricky as we cannot use `requireRule` due to circular dependencies. Instead the deprecated `BuildTargetPaths.getGenPath` was used to construct strings of paths from target names. Hopefully there is a better way to do it than this.

Reviewed By: milend

fbshipit-source-id: 182f51e35fcd7cb4f9a738e9692e82fa4f849722