Migrate SceneKitAssets to AbstractBuildRule
Summary:
Same as D25381726.

I wasn't able to reproduce this but some people told that buck doesn't track copy_scene_assets target dependency as well. I suppose it can be because of the same reason as with `libtool` and some other tools which we fixed previously.

Update:
What this rule does:
- If `copySceneAssets` tool is not present, it just copies all `sceneKitAssetsPaths` to `outputDir`.
- If it is present, it runs `copy-scene-assets <absolute input path> -o <outputDir>/<filename>` processing files one by one.

Both tool and list of inputs are marked with `AddToRuleKey`, so `depsSupplier` should have all required deps.

Reviewed By: cjhopman

fbshipit-source-id: 88fdb82319743743dfce0a262fc0f60a12494896