Be less strict on toolchain_target
Summary: Currently, if a toolchain_target is specified, no other fields may appear in the same block. This diff loosens the restriction if every other field that is not the toolchain_target is empty, this allows later rules to effectively "clear" config by making empty a sentinel state as well.

Reviewed By: cjhopman

fbshipit-source-id: ca7c189f681117140c801b9492003079d1448564