Make sure native libs and their symlinks are present at JavaTest runtime
Summary:
When we run tests with native libraries, we add symlinks pointing to the native libraries that we need. We need to make sure that these are present when we run, otherwise we'll get an error saying that these native libraries cannot be found.

To make this work with RE, we add them to the `requiredPaths` param.

Reviewed By: jiawei-lyu

fbshipit-source-id: e4c7a30d17671b575cf44069793fa460858c2d20