Don't require bootclasspath entries to be required paths
Summary:
Open to suggestions here, but this diff attempts to fix a bug when running Android robolectric tests with Remote Execution. In particular, before this diff, we will add the Android platform to the `requiredPaths`. That will mean that tpx can't normalise the paths correct, because the Android platform paths are not part of `buck-out` (they're in `/opt/android`). Additionally, the Android platform does not need to be copied onto the RE worker - it is already installed (since it is also needed when building Android libraries).

The fix is just to remove these entries from `requiredPaths`. We still add them to the classpath.

Reviewed By: jiawei-lyu

fbshipit-source-id: 151a711a632753c8bccc8f6da6c837d9c1f0180f