Add robolectric runtime dep dir as required paths only
Summary:
The way this dir works is that it contains a bunch of different versions of robolectric that can then be chosen at runtime. We don't want to include all of these jars on the classpath, as we don't want to package them up into the test jar. Instead, just make sure they are added as required paths so that the dir is present when we are running with RE.

This fixes D25510547

Reviewed By: jiawei-lyu

fbshipit-source-id: 4b3dc88257dcd3cdb32e9b07a24110b471da4118