Events posted at a specific timestamp populate only milliseconds timestamp
Summary:
Event's nano time is expected to be set to `System.nanoTime()` that is `the current value of the running Java Virtual Machine's high-resolution time source, in nanoseconds.
`

https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/xplat/build_infra/buck_client/src/com/facebook/buck/util/timing/Clock.java?commit=100352804f6844d7002a85cdf5a49b76001f52cf&lines=24-25

https://fburl.com/7d9co4gx -  `The value returned represents nanoseconds since some fixed but arbitrary origin time (perhaps in the future, so values may be negative).`

When we send an event with `Instant atTime` field (that actually designed for Downward API events) we can't calculate amount of nanos elapsed from some JVM's specific timestamp.

Reviewed By: irenewchen

fbshipit-source-id: 62345b3b0148c9f39a6b12d2e6d964a504e92b0b