Require android resource deps for Robolectric Tests if present
Summary:
When we're running with legacy resources we need to be able to access all the paths of those resources so that they can be parsed and put into a lookup table.

The way that these resources are built in Buck is via symlinks to the original file, so we need to follow the symlinks and make sure that those are required too.

Reviewed By: brianduff

fbshipit-source-id: 9644a1a153117f62c441dba7e06e08057f974441