UnusedDependenciesFinder doesn't need CellNameResolver
Summary:
As we use `CanonicalCellName` to write into a deps file, we could read from file and recreate `CanonicalCellName` without `CellNameResolver`.

Then we could get rid of `CellNameResolver` in `UnusedDependenciesFinder`.

This would help us to compact a bit protobuf command that we suppose to send between buck and javacd. (we would not need to pass known cells data)

Reviewed By: irenewchen

fbshipit-source-id: 0572b6e11730360e0440fe3270c5d9abc4c4fb6e