Change command failures to raise Exceptions
Summary:
After discussion with bobyangyf and ndmitchell, we are changing the behavior of buck_e2e so that a failure becomes a BuckException that is raised instead of a BuckResult that's returned.
Add a `BuckException` that extends both `Exception` and `BuckResult`
Also add a `expect_failure` function to `repo_workspace.py` that wraps around any failure buck command we expect in a test.

Reviewed By: bobyangyf

fbshipit-source-id: 9d7a0ef910919c7619437fce698a4aac23141765