HaskellIdeDescription
Summary:
## Context

I introduced the HaskellIdeRule last week as a fork of the Ghci rule in D25461671

## This stack

The goal is to be able to create haskell_ide targets in Starlark that produce a compilation database to load *multiple* Haskell projects from source in ghcide

## This diff

Introduce a description class for HaskellIdeRule, for use in Buck python/Starlark scripts. Given a set of deps (or a deps query), the description uses the `HaskellIdeDependencies` interface to collect IDE information for all the direct Haskell dependencies and construct a `HaskellIdeRule` that will produce compilation flags to load all of them *from source*.

Moreover, the constructed `HaskellIdeRule` will not actually depend on those Haskell dependencies, so that they will not be built. Only the non-Haskell direct dependencies and all the indirect dependencies will be built.

The description does not use the `srcs` argument for anything, but we validate that all of them are included in the direct Haskell deps.

Reviewed By: myszon

fbshipit-source-id: 5afddc82363edfbbbb72f3c74086c3d66cfe5dc5