HaskellSources.concat
Summary:
## Context

I introduced the HaskellIdeRule as a fork of the Ghci rule in order to support Haskell IDE functionality in D25461671

## This stack

The goal is to be able to create haskell_ide targets in Starlark that produce a compilation database to load *multiple* Haskell projects from source in ghcide

## This diff

The `concat` method allows to union multiple `HaskellSources` objects. This is needed to produce the list of source files to load in the IDE when working with multiple projects. The only tricky bit is checking for duplicate module names.

Reviewed By: myszon

fbshipit-source-id: 68449eff9f7a68c9f5320ed4c2281282097c30c0