Use expected size for hashset.
Summary:
Profiling showed that `getProjectElementForBuildTargets` was spending a bunch of time resizing the ImmutableSets it collects into. By default, ImmutableSet has an expected size of 16, but some modules have way more deps than this.

Calculate the exact size of the set in advance and allocate it accordingly. Doing this required converting the code in getProjectElementFromBuildTargets to not use stream / functional idioms, since it appears to be nontrivial to use a stream collector to an ImmutableSet with an expected size.

Reviewed By: cute-jumper

fbshipit-source-id: eb0926eef5ae1414a25c30ae9d9a07f298dee857