Improve ParsingJavaPackageFinder performance.
Summary:
PJPF has a "fast" and a "slow" strategy. The fast strategy is to read the first few lines of each java file, stopping when e.g. an "import" statement is reached, then pass those first few lines to be parsed as a Java file by the Eclipse java parser. If this fails, it falls back to the slow strategy which is parsing the whole file.

The fast strategy was failing 100% on a large codebase because the first few lines of the file were being smashed together without carriage returns, and because almost every .java file starts with a comment, this meant that the package statement was embedded in a comment, e.g.:

```
// (c) Foobar Inc.package com.facebook.blah.blah;
```

This is not only slow, but leads to variable performance, since the files chosen to represent each package are somewhat arbitrary, and some may take much longer to parse than others. There's a win on the average case, but an even bigger win on the p75 case because of the variability of this.

This diff retains the carriage returns, allowing the fast path to be used in all but 5 cases (all of which are files that have no package statement, e.g. empty files etc).

Reviewed By: cjhopman

fbshipit-source-id: 41341603397f2bd522303a683f9679c6ecfd75db