prepend / to debug prefix map entries
Summary:
Using debug prefix maps that replaced absolute paths with strings with no leading / would result in the debug info having the working directory prepended. For example, using `-fdebug-prefix-map=/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk=SDKROOT` we get in the debug info:

```
0x00000fa9:       DW_TAG_structure_type
                    DW_AT_name	("NSCoder")
                    DW_AT_linkage_name	("$sSo7NSCoderCD")
                    DW_AT_byte_size	(0x08)
                    DW_AT_decl_file	("/Users/me/source/SDKROOT/System/Library/Frameworks/Foundation.framework/Headers/NSCoder.h")
```

This results in debug info that is not possible to correctly remap. This change adds a leading / to all debug prefixes, so that the correct absolute path can be restored in the debugger.

Reviewed By: milend

fbshipit-source-id: 0af261e0213743db18ebc87c8255553d984e8c5b