add exposedPackages to rule keys
Summary:
# Context

I noticed that editing the dependencies of a target is not reflected in the IDE, not even after a restart, requiring a `buck clean`. This is because the rule keys of the `HaskellIdeRule` class are incomplete.

# This diff

`exposedPackages` has a direct effect on the output artifact and needs to be part of the rule key

Reviewed By: malmerey

fbshipit-source-id: 53c13b23b4020e87dd3779670efeb35fc3dcc1d8