Add explicitly specified targets to targets-under-test for code coverage
Summary:
Right now, if you call `buck test <JAVA_LIBRARY> --code-coverage`, and `JAVA_LIBRARY` has no `tests` attribute, then we won't record code-coverage for `JAVA_LIBRARY`. What we want is instead to record 0% code coverage for `JAVA_LIBRARY`. We can do that by adding `JAVA_LIBRARY` to the rules that we care about for code coverage.

This is put behind a config param so that they existing behavior is unchanged until the config option is enabled.

Reviewed By: brianduff

fbshipit-source-id: cd88fa7ec32d48337673a9ca21f3470606eca96d