relativize paths in argfile
Summary:
External test runners (tps and tpx) for instrumentation tests are
struggling to support argsfiles because the contents of it has to be updated
when the action is sent to Remote Execution (or run in sandbox).
The file contains paths to files, these files are already included to the
required_path list and are handled by tpx/tps correctly.
Buck will relativize all paths it puts into the argfile so external runner doesn't have to do anything else.

Reviewed By: mykola-semko

fbshipit-source-id: a52c333c00c35eb930f1055028136ac4b09a9dcc