do not pass swiftmodule includes to the ClangImporter
Summary: When building Swift we pass an include search path for dependent Swift libs swiftmodule files to the clang importer as well as a plain include to the Swift compiler. This is unnecessary, the clang importer does not understand swiftmodule files and the flag will be ignored. This diff removes the include from the preprocessor input.

Reviewed By: milend

fbshipit-source-id: 07c9af75219f48cf83e9269758eac925126a1960