Move java_for_tests_version to live with the java version it refers to
Summary:
Previously, we would be setting the java_for_tests_version in the `TestCommand`. That can actually be in a different cell to the cell we used when we were building the test, and so it can be inconsistent.

This diff makes it so that the two are tied together in the `JavaOptions`. Therefore, we should have a consistent value.

Reviewed By: jiawei-lyu

fbshipit-source-id: b1519a15335f6048936f4a37e3ee280ee2fcde16