Support per-rule hashed buck-out linking overrides
Summary:
Some binaries are built in a way that requires they be run in-place
(e.g. dynamically linked C++ binaries, in-place Python binaries) and
so aren't compatible with hashed buck-out hardlinking (as the
hardlinked executables e.g. can no longer find their shared libs
directory).

There's currently a way to make Buck use symlinking for everything
in a build, but no way to mix, depending on whether the individual
binaries support them or not.  This diff adds a way for descriptions to
opt-out of hardlinking on specific rules to allow using hardlinking
when possible.

Reviewed By: mzlee

fbshipit-source-id: 773f54e68832fb4db79fe9cf2f2158712cb0f75f